Rename rule and er modules.
