+++ src/Chromar.hs
@@ -1,13 +1,19 @@
 module Chromar
     ( module Chromar.Core
-    , module Chromar.Multiset
-    , module Chromar.RuleQuotesE
-    , module Chromar.RExprs
+    , module Chromar.Enriched.Parse
+    , module Chromar.Enriched.Syntax
+    , module Chromar.Enriched.TH
+    , module Chromar.Enriched.Zip
     , module Chromar.Experiment
+    , module Chromar.Multiset
+    , module Chromar.Rule.TH
     ) where
 
 import Chromar.Core
+import Chromar.Enriched.Parse
+import Chromar.Enriched.Syntax
+import Chromar.Enriched.TH
+import Chromar.Enriched.Zip
 import Chromar.Experiment
 import Chromar.Multiset
-import Chromar.RExprs
-import Chromar.RuleQuotesE
+import Chromar.Rule.TH
