+++ src/Chromar/Experiment.hs
@@ -2,7 +2,7 @@
 module Chromar.Experiment where
 
 import Chromar.Core
-import Chromar.RExprs (Er, at)
+import Chromar.Enriched.Syntax (Er, at)
 import qualified System.Random as R
 
 class ToSpaceSep a where
